+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
 - Support "Find implementations" of defmultis/defprotocols. #31
+- Fix commands, code actions not being applied after 2.0.0.
 
 ## 2.0.3