+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## 2.1.0
+
 - Support "Find implementations" of defmultis/defprotocols. #31
 - Fix commands, code actions not being applied after 2.0.0.
 - Improve "find declaration or usages" to show popup for references.